+++ bindings/nostr-ffi/src/nips/nip05.rs
@@ -13,19 +13,19 @@ use crate::nips::nip19::Nip19Profile;
 use crate::PublicKey;
 
 #[uniffi::export(default(proxy = None))]
-pub fn verify_nip05(public_key: &PublicKey, nip05: &str, proxy: Option<String>) -> Result<bool> {
+pub async fn verify_nip05(public_key: &PublicKey, nip05: &str, proxy: Option<String>) -> Result<bool> {
     let proxy: Option<SocketAddr> = match proxy {
         Some(proxy) => Some(proxy.parse()?),
         None => None,
     };
-    Ok(nip05::verify_blocking(public_key.deref(), nip05, proxy)?)
+    Ok(nip05::verify(public_key.deref(), nip05, proxy).await?)
 }
 
 #[uniffi::export(default(proxy = None))]
-pub fn get_nip05_profile(nip05: &str, proxy: Option<String>) -> Result<Arc<Nip19Profile>> {
+pub async fn get_nip05_profile(nip05: &str, proxy: Option<String>) -> Result<Arc<Nip19Profile>> {
     let proxy: Option<SocketAddr> = match proxy {
         Some(proxy) => Some(proxy.parse()?),
         None => None,
     };
-    Ok(Arc::new(nip05::get_profile_blocking(nip05, proxy)?.into()))
+    Ok(Arc::new(nip05::get_profile(nip05, proxy).await?.into()))
 }